+++ server/src/definitions/YmlDefinitionProvider.ts
@@ -1,6 +1,4 @@
-import { ParserRuleContext } from "antlr4ts";
 import { Definition, Location } from "vscode-languageserver";
-import { createLocation } from "../visitors/utils";
 import { IDefinitionLocation } from "./IDefinitionLocation";
 
 /**
@@ -21,7 +19,7 @@ export class YmlDefinitionProvider {
       return null;
     }
     const defs = this.definitions
-      .filter((defLoc) => defLoc.entityName.indexOf(entityName) >= 0)
+      .filter((defLoc) => defLoc.entityName === entityName)
       .map((defLoc) => defLoc.location)
       // Fill an array with the remaining locations.
       .reduce((prev: Location[], elem) => {